Rename home page component blocks.
