+++ components/home/PostsList.jsx
@@ -3,7 +3,7 @@ import Link from "next/link";
 import React from "react";
 import PostSmall from "../contentCards/PostSmall";
 
-export default function PostsHome({ posts }) {
+export default function PostsList({ posts }) {
   return (
     <>
       <div className="d-flex w-100 mb-4 align-items-center">
@@ -14,7 +14,8 @@ export default function PostsHome({ posts }) {
           data-bs-offset="0,5"
           data-bs-placement="left"
           title="View All Posts"
-          className="icon-btn ms-auto">
+          className="icon-btn ms-auto"
+        >
           <IconExternalLink size={18} />
         </Link>
       </div>
